fix(R.15): harden runtime heartbeat status cache
