+++ crates/atm-core/src/protocol.rs
@@ -231,6 +231,7 @@ pub struct TeamMemberHeartbeatResponse {
 #[serde(rename_all = "snake_case")]
 pub enum RuntimeMemberState {
     Unknown,
+    IdentityConflict,
     Offline,
     Idle,
     Active,
